APIs/azure.com/security-tasks/2015-06-01-preview/python-aiohttp-client/tests/test_tasks_controller.py
# coding: utf-8

import pytest
import json
from aiohttp import web

from openapi_server.models.security_task import SecurityTask
from openapi_server.models.security_task_list import SecurityTaskList
from openapi_server.models.tasks_list_by_home_region_default_response import TasksListByHomeRegionDefaultResponse


pytestmark = pytest.mark.asyncio

async def test_tasks_get_resource_group_level_task(client):
    """Test case for tasks_get_resource_group_level_task

    
    """
    params = [('api-version', 'api_version_example')]
    headers = { 
        'Accept': 'application/json',
        'Authorization': 'Bearer special-key',
    }
    response = await client.request(
        method='GET',
        path='/subscriptions/{subscription_id}/resourceGroups/{resource_group_name}/providers/Microsoft.Security/locations/{asc_location}/tasks/{task_name}'.format(subscription_id='subscription_id_example', resource_group_name='resource_group_name_example', asc_location='asc_location_example', task_name='task_name_example'),
        headers=headers,
        params=params,
        )
    assert response.status == 200, 'Response body is : ' + (await response.read()).decode('utf-8')


pytestmark = pytest.mark.asyncio

async def test_tasks_get_subscription_level_task(client):
    """Test case for tasks_get_subscription_level_task

    
    """
    params = [('api-version', 'api_version_example')]
    headers = { 
        'Accept': 'application/json',
        'Authorization': 'Bearer special-key',
    }
    response = await client.request(
        method='GET',
        path='/subscriptions/{subscription_id}/providers/Microsoft.Security/locations/{asc_location}/tasks/{task_name}'.format(subscription_id='subscription_id_example', asc_location='asc_location_example', task_name='task_name_example'),
        headers=headers,
        params=params,
        )
    assert response.status == 200, 'Response body is : ' + (await response.read()).decode('utf-8')


pytestmark = pytest.mark.asyncio

async def test_tasks_list(client):
    """Test case for tasks_list

    
    """
    params = [('api-version', 'api_version_example'),
                    ('$filter', 'filter_example')]
    headers = { 
        'Accept': 'application/json',
        'Authorization': 'Bearer special-key',
    }
    response = await client.request(
        method='GET',
        path='/subscriptions/{subscription_id}/providers/Microsoft.Security/tasks'.format(subscription_id='subscription_id_example'),
        headers=headers,
        params=params,
        )
    assert response.status == 200, 'Response body is : ' + (await response.read()).decode('utf-8')


pytestmark = pytest.mark.asyncio

async def test_tasks_list_by_home_region(client):
    """Test case for tasks_list_by_home_region

    
    """
    params = [('api-version', 'api_version_example'),
                    ('$filter', 'filter_example')]
    headers = { 
        'Accept': 'application/json',
        'Authorization': 'Bearer special-key',
    }
    response = await client.request(
        method='GET',
        path='/subscriptions/{subscription_id}/providers/Microsoft.Security/locations/{asc_location}/tasks'.format(subscription_id='subscription_id_example', asc_location='asc_location_example'),
        headers=headers,
        params=params,
        )
    assert response.status == 200, 'Response body is : ' + (await response.read()).decode('utf-8')


pytestmark = pytest.mark.asyncio

async def test_tasks_list_by_resource_group(client):
    """Test case for tasks_list_by_resource_group

    
    """
    params = [('api-version', 'api_version_example'),
                    ('$filter', 'filter_example')]
    headers = { 
        'Accept': 'application/json',
        'Authorization': 'Bearer special-key',
    }
    response = await client.request(
        method='GET',
        path='/subscriptions/{subscription_id}/resourceGroups/{resource_group_name}/providers/Microsoft.Security/locations/{asc_location}/tasks'.format(subscription_id='subscription_id_example', resource_group_name='resource_group_name_example', asc_location='asc_location_example'),
        headers=headers,
        params=params,
        )
    assert response.status == 200, 'Response body is : ' + (await response.read()).decode('utf-8')


pytestmark = pytest.mark.asyncio

async def test_tasks_update_resource_group_level_task_state(client):
    """Test case for tasks_update_resource_group_level_task_state

    
    """
    params = [('api-version', 'api_version_example')]
    headers = { 
        'Accept': 'application/json',
        'Authorization': 'Bearer special-key',
    }
    response = await client.request(
        method='POST',
        path='/subscriptions/{subscription_id}/resourceGroups/{resource_group_name}/providers/Microsoft.Security/locations/{asc_location}/tasks/{task_name}/{task_update_action_type}'.format(subscription_id='subscription_id_example', resource_group_name='resource_group_name_example', asc_location='asc_location_example', task_name='task_name_example', task_update_action_type='task_update_action_type_example'),
        headers=headers,
        params=params,
        )
    assert response.status == 200, 'Response body is : ' + (await response.read()).decode('utf-8')


pytestmark = pytest.mark.asyncio

async def test_tasks_update_subscription_level_task_state(client):
    """Test case for tasks_update_subscription_level_task_state

    
    """
    params = [('api-version', 'api_version_example')]
    headers = { 
        'Accept': 'application/json',
        'Authorization': 'Bearer special-key',
    }
    response = await client.request(
        method='POST',
        path='/subscriptions/{subscription_id}/providers/Microsoft.Security/locations/{asc_location}/tasks/{task_name}/{task_update_action_type}'.format(subscription_id='subscription_id_example', asc_location='asc_location_example', task_name='task_name_example', task_update_action_type='task_update_action_type_example'),
        headers=headers,
        params=params,
        )
    assert response.status == 200, 'Response body is : ' + (await response.read()).decode('utf-8')

